Commit Of README files for General/Generic
We needed more general product and SDK top level descriptions of why to use TDI and why to use the Java SDK in the first place to give new developers of the platform and more high level overview.

It makes the SDK documentation more complete to have documentation around this.

some of the terms might need some adjustment in the future but this might be good enough for a first pass.